Fix some instances of -Wsign-compare and -Wignored-qualifiers
Fix instances in chrome/browser/ash/ as a precursor to reenable those
warnings in //third_party/shell-encryption.

Bug: 1095122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3921579
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1052317}

+++ chrome/browser/ash/arc/accessibility/ax_tree_source_arc.cc
@@ -594,8 +594,9 @@ void AXTreeSourceArc::GetChildren(
       continue;
 
     // move the current node to be visited later if necessary.
-    for (int j = i; j + 1 < out_children->size() &&
-                    NeedReorder(out_children->at(j), out_children->at(j + 1));
+    for (size_t j = i;
+         j + 1 < out_children->size() &&
+         NeedReorder(out_children->at(j), out_children->at(j + 1));
          j++) {
       std::swap(out_children->at(j), out_children->at(j + 1));
     }
