+++ chrome/browser/ash/file_manager/copy_or_move_io_task_impl.h
@@ -120,7 +120,7 @@ class CopyOrMoveIOTaskImpl {
   bool completed_ = false;
 
   // The number of file for which the file size has been fetched.
-  int files_preprocessed_ = 0;
+  size_t files_preprocessed_ = 0;
 
   // Stores the size of each source so we know what to increment the progress
   // bytes by for each copy or move completion.
